+++ functions/src/notifications/updateDaysFromFirstDay.ts
@@ -3,6 +3,7 @@
 import { onDocumentCreated } from 'firebase-functions/v2/firestore';
 import admin from '../admin';
 import { Board } from '../types/Board';
+import { isWorkingDay } from './isWorkingDay';
 
 export const updatePostDaysFromFirstDay = onDocumentCreated('/boards/{boardId}/posts/{postId}', async (event) => {
   const postId = event.params.postId;
@@ -53,28 +54,4 @@ function calculateWorkingDaysFromFirstDay(firstDay: Date): number {
   const workingDaysCount = daysArray.filter(isWorkingDay).length;
 
   return workingDaysCount;
-}
-
-// Function takes date and return if is working day
-function isWorkingDay(date: Date): boolean {
-  // Exclude Sundays (0) and Saturdays (6)
-  if (date.getDay() === 0 || date.getDay() === 6) {
-    return false;
-  }
-
-  const temporaryHolidays: Record<string, boolean> = {
-    '2024-12-31': true, // 신정
-    '2025-01-01': true, // 신정
-    '2025-01-28': true, // 설날
-    '2025-01-29': true, // 설날
-    '2025-01-30': true, // 설날
-  };
-
-  const yearMonthDay = date.toISOString().split('T')[0];
-
-  if (temporaryHolidays[yearMonthDay]) {
-    return false;
-  }
-
-  return true;
 }
